+++ src/web/assets/src/js/ImageOptimize.js
@@ -1,3 +1,2 @@
 // Import our CSS
 import '@/css/app.pcss';
-import '@/img/ImageOptimize-icon.svg';